+++ packages/humanink-storage/src/index.ts
@@ -1,1 +1,4 @@
-﻿export { FileContentRepository } from './file-content-repository.js';
+export { FileContentRepository } from './file-content-repository.js';
+
+export * from './workbench/file-workbench-repository.js';
+export * from './workbench/overlay-store.js';